+++ lib/apexcharts/charts/radar.rb
@@ -1,0 +1,10 @@
+module Apexcharts
+  class RadarChart < CartesianChart
+    def chart_type
+      'radar'
+    end
+
+    def mixed_series
+    end
+  end
+end